+++ common/src/main/java/com/hyfata/najoan/koreanpatch/util/mixin/editbox/IEditBoxAccessor.java
@@ -1,4 +1,4 @@
-package com.hyfata.najoan.koreanpatch.util.mixin.textfieldwidget;
+package com.hyfata.najoan.koreanpatch.util.mixin.editbox;
 
 import java.util.function.Consumer;
