+++ fabric/src/main/java/com/hyfata/najoan/koreanpatch/mixin/mods/bettercommand/MultiLineEditBoxMixin.java
@@ -3,8 +3,8 @@ package com.hyfata.najoan.koreanpatch.mixin.mods.bettercommand;
 import bettercommandblockui.main.ui.MultiLineTextFieldWidget;
 import com.hyfata.najoan.koreanpatch.client.KoreanPatchClient;
 import com.hyfata.najoan.koreanpatch.util.language.LanguageUtil;
-import com.hyfata.najoan.koreanpatch.util.mixin.textfieldwidget.IEditBoxAccessor;
-import com.hyfata.najoan.koreanpatch.util.mixin.textfieldwidget.EditBoxHandler;
+import com.hyfata.najoan.koreanpatch.util.mixin.editbox.IEditBoxAccessor;
+import com.hyfata.najoan.koreanpatch.util.mixin.editbox.EditBoxHandler;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.components.EditBox;
